moved ListVersions API view to separate file
